wake: fix crash on undef variable in top scope

+++ expr.cpp
@@ -61,7 +61,7 @@ static void format(std::ostream &os, int depth, const Expr *expr) {
   } else if (expr->type == DefBinding::type) {
     const DefBinding *def = reinterpret_cast<const DefBinding*>(expr);
     os << pad(depth) << "DefBinding @ " << def->location.str() << std::endl;
-    size_t vals = (int)def->val.size();
+    int vals = def->val.size();
     for (auto &i : def->order) {
       os << pad(depth+2) << (i.second < vals ? "val " : "fun ") << i.first << " =" << std::endl;
       format(os, depth+4, (i.second < vals) ? def->val[i.second].get() : def->fun[i.second - vals].get());
